requirements.txt
click==8.0.3
    # via -r requirements.in
prompt-toolkit==3.0.21
    # via -r requirements.in
togglpy==0.1.2
    # via -r requirements.in
wcwidth==0.2.5
    # via prompt-toolkit
